refactor: migrate TeachersUnite to TS
