+++ src/pages/TeachersUnite/KnowATeacherPage.tsx
@@ -1,10 +1,8 @@
 import Str from 'expensify-common/lib/str';
-import lodashGet from 'lodash/get';
-import PropTypes from 'prop-types';
 import React, {useCallback} from 'react';
 import {View} from 'react-native';
 import {withOnyx} from 'react-native-onyx';
-import _ from 'underscore';
+import type {OnyxEntry} from 'react-native-onyx/lib/types';
 import FormProvider from '@components/Form/FormProvider';
 import InputWrapper from '@components/Form/InputWrapper';
 import HeaderWithBackButton from '@components/HeaderWithBackButton';
@@ -22,32 +20,29 @@ import TeachersUnite from '@userActions/TeachersUnite';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import ROUTES from '@src/ROUTES';
+import type {LoginList} from '@src/types/onyx';
 
-const propTypes = {
-    /** Login list for the user that is signed in */
-    loginList: PropTypes.shape({
-        /** Phone/Email associated with user */
-        partnerUserID: PropTypes.string,
-    }),
+type KnowATeacherFormData = {
+    firstName: string;
+    lastName: string;
+    partnerUserID: string;
 };
 
-const defaultProps = {
-    loginList: {},
+type KnowATeacherPageOnyxProps = {
+    loginList: OnyxEntry<LoginList>;
 };
 
-function KnowATeacherPage(props) {
+type KnowATeacherPageProps = KnowATeacherPageOnyxProps;
+
+function KnowATeacherPage(props: KnowATeacherPageProps) {
     const styles = useThemeStyles();
     const {translate} = useLocalize();
     const {isProduction} = useEnvironment();
 
     /**
      * Submit form to pass firstName, partnerUserID and lastName
-     * @param {Object} values
-     * @param {String} values.partnerUserID
-     * @param {String} values.firstName
-     * @param {String} values.lastName
      */
-    const onSubmit = (values) => {
+    const onSubmit = (values: KnowATeacherFormData) => {
         const phoneLogin = LoginUtils.getPhoneLogin(values.partnerUserID);
         const validateIfnumber = LoginUtils.validateNumber(phoneLogin);
         const contactMethod = (validateIfnumber || values.partnerUserID).trim().toLowerCase();
@@ -66,34 +61,34 @@ function KnowATeacherPage(props) {
      * @returns {Object} - An object containing the errors for each inputID
      */
     const validate = useCallback(
-        (values) => {
+        (values: KnowATeacherFormData) => {
             const errors = {};
             const phoneLogin = LoginUtils.getPhoneLogin(values.partnerUserID);
             const validateIfnumber = LoginUtils.validateNumber(phoneLogin);
 
             if (!ValidationUtils.isValidLegalName(values.firstName)) {
-                ErrorUtils.addErrorMessage(errors, 'firstName', translate('privatePersonalDetails.error.hasInvalidCharacter'));
-            } else if (_.isEmpty(values.firstName)) {
-                ErrorUtils.addErrorMessage(errors, 'firstName', translate('bankAccount.error.firstName'));
+                ErrorUtils.addErrorMessage(errors, 'firstName', 'privatePersonalDetails.error.hasInvalidCharacter');
+            } else if (values.firstName) {
+                ErrorUtils.addErrorMessage(errors, 'firstName', 'bankAccount.error.firstName');
             }
             if (!ValidationUtils.isValidLegalName(values.lastName)) {
-                ErrorUtils.addErrorMessage(errors, 'lastName', translate('privatePersonalDetails.error.hasInvalidCharacter'));
-            } else if (_.isEmpty(values.lastName)) {
-                ErrorUtils.addErrorMessage(errors, 'lastName', translate('bankAccount.error.lastName'));
+                ErrorUtils.addErrorMessage(errors, 'lastName', 'privatePersonalDetails.error.hasInvalidCharacter');
+            } else if (values.lastName) {
+                ErrorUtils.addErrorMessage(errors, 'lastName', 'bankAccount.error.lastName');
             }
-            if (_.isEmpty(values.partnerUserID)) {
-                ErrorUtils.addErrorMessage(errors, 'partnerUserID', translate('teachersUnitePage.error.enterPhoneEmail'));
+            if (values.partnerUserID) {
+                ErrorUtils.addErrorMessage(errors, 'partnerUserID', 'teachersUnitePage.error.enterPhoneEmail');
             }
-            if (!_.isEmpty(values.partnerUserID) && lodashGet(props.loginList, validateIfnumber || values.partnerUserID.toLowerCase())) {
+            if (!values.partnerUserID && props.loginList?.[validateIfnumber || values.partnerUserID.toLowerCase()]) {
                 ErrorUtils.addErrorMessage(errors, 'partnerUserID', 'teachersUnitePage.error.tryDifferentEmail');
             }
-            if (!_.isEmpty(values.partnerUserID) && !(validateIfnumber || Str.isValidEmail(values.partnerUserID))) {
+            if (!values.partnerUserID && !(validateIfnumber || Str.isValidEmail(values.partnerUserID))) {
                 ErrorUtils.addErrorMessage(errors, 'partnerUserID', 'contacts.genericFailureMessages.invalidContactMethod');
             }
 
             return errors;
         },
-        [props.loginList, translate],
+        [props.loginList],
     );
 
     return (
@@ -105,6 +100,7 @@ function KnowATeacherPage(props) {
                 title={translate('teachersUnitePage.iKnowATeacher')}
                 onBackButtonPress={() => Navigation.goBack(ROUTES.TEACHERS_UNITE)}
             />
+            {/* @ts-expect-error TODO: Remove this once FormProvider (https://github.com/Expensify/App/issues/31972) is migrated to TypeScript. */}
             <FormProvider
                 enabledWhenOffline
                 style={[styles.flexGrow1, styles.ph5]}
@@ -116,6 +112,7 @@ function KnowATeacherPage(props) {
                 <Text style={[styles.mb6]}>{translate('teachersUnitePage.getInTouch')}</Text>
                 <View>
                     <InputWrapper
+                        // @ts-expect-error TODO: Remove this once InputWrapper (https://github.com/Expensify/App/issues/31972) is migrated to TypeScript.
                         InputComponent={TextInput}
                         inputID="firstName"
                         name="fname"
@@ -128,6 +125,7 @@ function KnowATeacherPage(props) {
                 </View>
                 <View style={styles.mv4}>
                     <InputWrapper
+                        // @ts-expect-error TODO: Remove this once InputWrapper (https://github.com/Expensify/App/issues/31972) is migrated to TypeScript.
                         InputComponent={TextInput}
                         inputID="lastName"
                         name="lname"
@@ -140,6 +138,7 @@ function KnowATeacherPage(props) {
                 </View>
                 <View>
                     <InputWrapper
+                        // @ts-expect-error TODO: Remove this once InputWrapper (https://github.com/Expensify/App/issues/31972) is migrated to TypeScript.
                         InputComponent={TextInput}
                         inputID="partnerUserID"
                         name="partnerUserID"
@@ -155,10 +154,8 @@ function KnowATeacherPage(props) {
     );
 }
 
-KnowATeacherPage.propTypes = propTypes;
-KnowATeacherPage.defaultProps = defaultProps;
 KnowATeacherPage.displayName = 'KnowATeacherPage';
 
-export default withOnyx({
+export default withOnyx<KnowATeacherPageProps, KnowATeacherPageOnyxProps>({
     loginList: {key: ONYXKEYS.LOGIN_LIST},
 })(KnowATeacherPage);
